requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
bce-python-sdk==0.9.22
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dashscope==1.20.10
dataclasses-json==0.6.7
datasets==2.20.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
emoji==2.13.2
eventlet==0.37.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
filelock==3.15.3
Flask==3.0.3
Flask-SocketIO==5.3.7
Flask-WTF==1.2.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.5
llama-cloud==0.1.0
llama-index==0.11.12
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.12
llama-index-embeddings-dashscope==0.2.1
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post3
llama-index-llms-dashscope==0.2.1
llama-index-llms-ollama==0.3.2
llama-index-llms-openai==0.2.9
llama-index-llms-openai-like==0.2.0
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.6
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
minijinja==2.0.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
ollama==0.3.3
openai==1.47.1
packaging==24.1
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
pypinyin==0.51.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-pptx==0.6.23
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.1
qianfan==0.4.8.1
regex==2024.5.15
requests==2.32.3
rich==13.8.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
striprtf==0.0.26
sympy==1.13.0
tenacity==8.4.1
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zhipuai==2.1.5.20230904
zipp==3.19.2
